% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rfunc.R
\name{save_rfunc}
\alias{save_rfunc}
\title{Save the function f for calling from fortran}
\usage{
save_rfunc(f)
}
\arguments{
\item{f}{the R function to be called using \code{.Fortran}}
}
\description{
Save the function f for calling from fortran
}
